+++ R/clearBedCache.R
@@ -17,6 +17,8 @@ clearBedCache <- function(
    if(!checkBedConn(verbose=FALSE)){
       stop("Unsuccessful connection")
    }
+   ## Write cache in the user file space only if the "useCache" parameter
+   ## is set to TRUE when calling `connectToBed()` (default: useCache=FALSE)
    if(!get("useCache", bedEnv)){
       warning("Cache is OFF: nothing is done")
       invisible(NULL)