+++ R/dumpUniprotDb.R
@@ -5,11 +5,13 @@
 #' @param taxOfInt the organism of interest. Only human ("9606"),
 #' mouse ("10090") and rat ("10116") are supported
 #' @param release the release of interest (check if already downloaded)
+#' @param ddir path to the directory where the data should be saved
 #' @param env the R environment in which to load the tables when built
 #'
 dumpUniprotDb <- function(
     taxOfInt,
     release,
+    ddir,
     env=parent.frame(n=1)
 ){
 
@@ -30,7 +32,7 @@ dumpUniprotDb <- function(
 
     ## Download files if necessary ----
     ftp <- "ftp://ftp.uniprot.org/pub/databases/uniprot/current_release/knowledgebase/taxonomic_divisions"
-    dumpDir <- "Uniprot-DATA"
+    dumpDir <- file.path(ddir, "Uniprot-DATA")
     if(file.exists(dumpDir)){
         load(file.path(dumpDir, "dumpRelease.rda"))
         message("Last release: ", dumpRelease)