+++ R/getEnsemblGeneIds.R
@@ -6,6 +6,7 @@
 #' of interest (e.g. "Homo sapiens")
 #' @param release the Ensembl release of interest (e.g. "83")
 #' @param gv the genome version (e.g. "38")
+#' @param ddir path to the directory where the data should be saved
 #' @param dbCref a named vector of characters providing cross-reference DB of
 #' interest. These DB are also used to find indirect ID associations.
 #' @param dbAss a named vector of characters providing associated DB of
@@ -18,6 +19,7 @@ getEnsemblGeneIds <- function(
    organism,
    release,
    gv,
+   ddir,
    dbCref,
    dbAss,
    canChromosomes
@@ -38,7 +40,7 @@ getEnsemblGeneIds <- function(
       "stable_id_event", "mapping_session",
       "seq_region", "coord_system"
    )
-   dumpEnsCore(organism, release, gv, toDump)
+   dumpEnsCore(organism, release, gv, ddir, toDump)
 
    ################################################@
    ## Current Ensembl database and organism ----