+++ R/zzz.R
@@ -13,10 +13,10 @@
 #' All the connections that have been saved can be listed
 #' with [lsBedConnections] and any of
 #' them can be forgotten with [forgetBedConnection].
-#' @param useCache if TRUE the results of large queries can be saved localy
+#' @param useCache if TRUE the results of large queries can be saved locally
 #' in a file. The default is FALSE for policy reasons.
 #' But it is recommended to set it to TRUE to improve the speed
-#' of recurent queries.
+#' of recurrent queries.
 #' If NA (default parameter) the value is taken from former connection if
 #' it exists or it is set to FALSE.
 #' @param importPath the path to the import folder for loading information
@@ -25,11 +25,11 @@
 #' @return This function does not return any value. It prepares the BED
 #' environment to allow transparent DB calls.
 #'
-#' @details Be carefull that you should reconnect to BED database each time
+#' @details Be careful that you should reconnect to BED database each time
 #' the environment is reloaded. It is done automatically if `remember` is
 #' set to TRUE.
 #'
-#' Information about how to get an instance of the BED neo4j database is
+#' Information about how to get an instance of the BED 'Neo4j' database is
 #' provided here:
 #' - <https://github.com/patzaw/BED#bed-database-instance-available-as-a-docker-image>
 #' - <https://github.com/patzaw/BED#build-a-bed-database-instance>
@@ -184,12 +184,15 @@ lsBedConnections <- function(){
 #' Forget a BED connection
 #'
 #' @param connection the id of the connection to forget.
+#' @param save a logical. Should be set to TRUE to save the updated list of
+#' connections in the file space (default to FALSE to comply with CRAN
+#' policies).
 #'
 #' @seealso [lsBedConnections], [checkBedConn], [connectToBed]
 #'
 #' @export
 #'
-forgetBedConnection <- function(connection){
+forgetBedConnection <- function(connection, save=FALSE){
    conFile <- file.path(
       Sys.getenv("HOME"), "R", "BED", "BED-Connections.rda"
    )
@@ -198,7 +201,9 @@ forgetBedConnection <- function(connection){
       load(conFile)
    }
    connections <- connections[-connection]
-   save(connections, file=conFile)
+   if(save){
+      save(connections, file=conFile)
+   }
 }
 
 ###############################################################################@